Add rotation and destination to the shrimp
