+++ Assets/Prefabs/Shrimp Enemy.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b5667a139d9e68740aba6ac9a240c400
+guid: 7df56ecc5b990c04db208c3d008f1508
 PrefabImporter:
   externalObjects: {}
   userData:
