+++ Assets/Scenes/Enemy Shrimp.unity
@@ -260,37 +260,6 @@ MonoBehaviour:
     mipBias: 0
     varianceClampScale: 0.9
     contrastAdaptiveSharpening: 0
---- !u!1 &657598536
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 657598537}
-  m_Layer: 0
-  m_Name: Destination
-  m_TagString: Untagged
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 1
---- !u!4 &657598537
-Transform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 657598536}
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 1.72, y: -0.39959186, z: -0.016986646}
-  m_LocalScale: {x: 1, y: 1, z: 1}
-  m_ConstrainProportionsScale: 0
-  m_Children: []
-  m_Father: {fileID: 0}
-  m_RootOrder: 3
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &1595397856
 GameObject:
   m_ObjectHideFlags: 0
@@ -439,7 +408,7 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 1
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!1001 &8556755169351438298
+--- !u!1001 &2107491858347714079
 PrefabInstance:
   m_ObjectHideFlags: 0
   serializedVersion: 2
@@ -447,60 +416,61 @@ PrefabInstance:
     serializedVersion: 3
     m_TransformParent: {fileID: 0}
     m_Modifications:
-    - target: {fileID: 4948201144338891138, guid: b5667a139d9e68740aba6ac9a240c400, type: 3}
+    - target: {fileID: 1682477695393711541, guid: 7df56ecc5b990c04db208c3d008f1508, type: 3}
+      propertyPath: m_Name
+      value: Shrimp Enemy
+      objectReference: {fileID: 0}
+    - target: {fileID: 3150888720566314562, guid: 7df56ecc5b990c04db208c3d008f1508, type: 3}
       propertyPath: m_RootOrder
       value: 2
       objectReference: {fileID: 0}
-    - target: {fileID: 4948201144338891138, guid: b5667a139d9e68740aba6ac9a240c400, type: 3}
+    - target: {fileID: 3150888720566314562, guid: 7df56ecc5b990c04db208c3d008f1508, type: 3}
       propertyPath: m_LocalPosition.x
-      value: -2.22
+      value: -0.45769417
       objectReference: {fileID: 0}
-    - target: {fileID: 4948201144338891138, guid: b5667a139d9e68740aba6ac9a240c400, type: 3}
+    - target: {fileID: 3150888720566314562, guid: 7df56ecc5b990c04db208c3d008f1508, type: 3}
       propertyPath: m_LocalPosition.y
-      value: -0.38
+      value: -0.31506506
       objectReference: {fileID: 0}
-    - target: {fileID: 4948201144338891138, guid: b5667a139d9e68740aba6ac9a240c400, type: 3}
+    - target: {fileID: 3150888720566314562, guid: 7df56ecc5b990c04db208c3d008f1508, type: 3}
       propertyPath: m_LocalPosition.z
-      value: 0
+      value: -0.032870274
       objectReference: {fileID: 0}
-    - target: {fileID: 4948201144338891138, guid: b5667a139d9e68740aba6ac9a240c400, type: 3}
+    - target: {fileID: 3150888720566314562, guid: 7df56ecc5b990c04db208c3d008f1508, type: 3}
       propertyPath: m_LocalRotation.w
       value: 1
       objectReference: {fileID: 0}
-    - target: {fileID: 4948201144338891138, guid: b5667a139d9e68740aba6ac9a240c400, type: 3}
+    - target: {fileID: 3150888720566314562, guid: 7df56ecc5b990c04db208c3d008f1508, type: 3}
       propertyPath: m_LocalRotation.x
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 4948201144338891138, guid: b5667a139d9e68740aba6ac9a240c400, type: 3}
+    - target: {fileID: 3150888720566314562, guid: 7df56ecc5b990c04db208c3d008f1508, type: 3}
       propertyPath: m_LocalRotation.y
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 4948201144338891138, guid: b5667a139d9e68740aba6ac9a240c400, type: 3}
+    - target: {fileID: 3150888720566314562, guid: 7df56ecc5b990c04db208c3d008f1508, type: 3}
       propertyPath: m_LocalRotation.z
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 4948201144338891138, guid: b5667a139d9e68740aba6ac9a240c400, type: 3}
+    - target: {fileID: 3150888720566314562, guid: 7df56ecc5b990c04db208c3d008f1508, type: 3}
       propertyPath: m_LocalEulerAnglesHint.x
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 4948201144338891138, guid: b5667a139d9e68740aba6ac9a240c400, type: 3}
+    - target: {fileID: 3150888720566314562, guid: 7df56ecc5b990c04db208c3d008f1508, type: 3}
       propertyPath: m_LocalEulerAnglesHint.y
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 4948201144338891138, guid: b5667a139d9e68740aba6ac9a240c400, type: 3}
+    - target: {fileID: 3150888720566314562, guid: 7df56ecc5b990c04db208c3d008f1508, type: 3}
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 5757247812233192999, guid: b5667a139d9e68740aba6ac9a240c400, type: 3}
-      propertyPath: Destination
-      value: 
-      objectReference: {fileID: 657598537}
-    - target: {fileID: 8717759906087337493, guid: b5667a139d9e68740aba6ac9a240c400, type: 3}
-      propertyPath: m_Name
-      value: Shrimp
+    - target: {fileID: 4300063104499277222, guid: 7df56ecc5b990c04db208c3d008f1508, type: 3}
+      propertyPath: m_Enabled
+      value: 1
       objectReference: {fileID: 0}
     m_RemovedComponents: []
-    m_RemovedGameObjects: []
+    m_RemovedGameObjects:
+    - {fileID: 8290509401930205160, guid: 7df56ecc5b990c04db208c3d008f1508, type: 3}
     m_AddedGameObjects: []
     m_AddedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: b5667a139d9e68740aba6ac9a240c400, type: 3}
+  m_SourcePrefab: {fileID: 100100000, guid: 7df56ecc5b990c04db208c3d008f1508, type: 3}
