+++ Assets/Scripts/Enemies.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6fea9db0da279464f9fe8b09282e1b81
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
