+++ Assets/Scripts/Enemies/Shrimp.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a58fc361f037fb34e96f2d8ccc177abf
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
