+++ Assets/Scripts/HideOnPlay.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d42c0d193319dd442a19c8aa0d0128d3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
